+++ js/ngRiffle/docs/changelog.md
@@ -1,3 +1,12 @@
+START__0.4.7
 ## Changes in v0.4.7
+
+END__0.4.7
+START__0.4.8
+## Changes in v0.4.8
+* [Storage API](#RiffleStorage)
+* Bug fix where rejecting a promise that is being returned via a call results in the success handler being called on the receiving end.
+
+END__0.4.8
 
 ## Changes in upcoming versions